requirements.txt
# AI Operations Assistant Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM Integration
google-generativeai>=0.8.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# CLI Interface
typer>=0.9.0
rich>=13.7.0

# Web UI
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
cachetools>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
